requirements.txt
django==2.1.4
graphene-django==2.2.0
django-filter==2.0.0
django-graphql-jwt==0.1.5
websocket==0.2.1
websocket-client==1.4.1